refactor: move `EVMEmptyExec` and `EVMTableExec` to own file
